+++ SwiftyMenu/Classes/MenuAttributes/SwiftyMenuAttributes+Accessory.swift
@@ -1,7 +1,7 @@
 //
-//  SwiftMenuDisplayable.swift
+//  SwiftyMenuAttributes+Accessory.swift
 //
-//  Copyright (c) 2019-2020 Karim Ebrahem (https://twitter.com/k_ebrahem_)
+//  Copyright (c) 2019-2021 Karim Ebrahem (https://twitter.com/k_ebrahem_)
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -24,11 +24,25 @@
 
 import Foundation
 
-/// `SwiftyMenuDisplayable` is the markable interface to allow custom types to be used with SwiftyMenu.
-public protocol SwiftyMenuDisplayable {
-    /// The value that will be displayed in the menu item
-    var displayableValue: String { get }
-    
-    /// The value that will be returned when select menu item
-    var retrievableValue: Any { get }
+public extension SwiftyMenuAttributes {
+
+    /** Describes the event of scroll user interaction */
+    enum Accessory {
+
+        /** The scroll ability is totally disabled */
+        case disabled
+
+        /** The scroll abiliby is enabled */
+        case enabled
+
+        var isEnabled: Bool {
+            switch self {
+            case .disabled:
+                return false
+            default:
+                return true
+            }
+        }
+    }
 }
+