+++ SwiftyMenu/Classes/SwiftyMenuAnimationStyle.swift
@@ -1,5 +1,5 @@
 //
-//  SwiftMenuDisplayable.swift
+//  SwiftyMenuAnimationStyle.swift
 //
 //  Copyright (c) 2019-2020 Karim Ebrahem (https://twitter.com/k_ebrahem_)
 //
@@ -24,11 +24,22 @@
 
 import Foundation
 
-/// `SwiftyMenuDisplayable` is the markable interface to allow custom types to be used with SwiftyMenu.
-public protocol SwiftyMenuDisplayable {
-    /// The value that will be displayed in the menu item
-    var displayableValue: String { get }
+/// Type describing the Animation Styles used to create the animation of Expanding and Collapsing `SwiftyMenu`.
+public enum AnimationStyle {
+    /// `SwiftyMenu` animation should be linear (Smooth Animation)
+    case linear
     
-    /// The value that will be returned when select menu item
-    var retrievableValue: Any { get }
+    /// `SwiftyMenu` animation should be spring (Bouncy Animation)
+    case spring(level: SpringPowerLevel)
+    
+    /// Defines how bouncy the animation should be
+    ///
+    /// - low: Bit of smooth and a bit of bounciness at the end
+    /// - normal: Not too bouncy and not too smooth
+    /// - high: Too bouncy
+    public enum SpringPowerLevel: Double {
+        case low = 0.75
+        case normal = 1.0
+        case high = 1.5
+    }
 }