+++ CHANGELOG.md
@@ -4,56 +4,103 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2021-07-04
+### Added
+- SwiftMenu different Attributes.
+- Support for SPM.
+
+### Changed
+- Removed all `IBDesignable` from SwiftyMenu and changed the styling of the menu using `SwiftyMenuAttributes`.
+
+### Fixed
+- Arrow direction when multi selection enabled.
+
+
+## [0.6.5] - 2021-03-15
+### Fixed
+- UI issues.
+
+## [0.6.4] - 2020-07-13
+### Changed
+- Update SnapKit version.
+
+### Fixed
+- Support RTL in SwiftyMenu.
+
+## [0.6.3] - 2020-04-06
+### Added
+- Allow Changing separator color.
+- Allow Changing separator characters.
+
+## [0.6.2] - 2020-04-03
+### Fixed
+- Fix arrow rotation when open or close SwiftyMenu
+
+## [0.6.1] - 2019-03-25
+### Added
+- Add Unit Tests to SwiftMenu with Coverage of 57%.
+
+## [0.6.0] - 2019-03-25
+### Added
+- Restucture project code and add some Docs.
+- Allow to add SwiftyMenu programmatically.
+- Allow to toggle SwiftyMenu from Code.
+
+## [0.5.9] - 2019-02-20
+### Changed
+- SwiftyMenu is now can adapt on number of items it contain when the scrolling is enabled.
+- Remove support for Swift 4 and only focus on Swift 5 and above.
+
 ## [0.5.8] - 2019-09-03
-**Fixed**
+### Fixed
 - Fix pre selected index in single selection didn't update label.
 
 ## [0.5.7] - 2019-09-03
-**Fixed**
+### Fixed
 - Fix pre selected index in single selection didn't update label.
 
 ## [0.5.6] - 2019-09-02
-**Fixed**
+### Fixed
 - Fix pre selected indices didn't update label.
 
 ## [0.5.5] - 2019-09-02
-**Fixed**
+### Fixed
 - Fix pre selected indices didn't update label.
 
 ## [0.5.4] - 2019-09-02
-**Fixed**
+### Fixed
 - Delete extra comma at the end when menu is multi selection.
 - Fix title label truncate.
 
 ## [0.5.3] - 2019-09-02
-**Fixed**
+### Fixed
 - Delete extra comma at the end when menu is multi selection.
 - Fix title label truncate.
 
 ## [0.5.2] - 2019-06-17
-**Fixed**
+### Fixed
 - Delete un expected animation while select option.
 - Fix arrow space in different sizes.
 
 ## [0.5.1] - 2019-06-17
-**Fixed**
+### Fixed
 - Delete un expected animation while select option.
 - Fix arrow space in different sizes.
 
 ## [0.5.0] - 2019-06-01
-**Added**
+### Added
 - Support to generic datasource.
 
 ## [0.4.9] - 2019-06-01
-**Added**
+### Added
 - Support to generic datasource.
 
 ## [0.4.8] - 2019-05-06
-**Added**
+### Added
 - Automate release new version to Cocoapods from Travis CI.
 - Add CHANGELOG file for the project.
 
 ## [0.4.7] - 2019-05-06
-**Added**
+### Added
 - Automate release new version to Cocoapods from Travis CI.
 - Add CHANGELOG file for the project.